requirements.txt
streamlit
ARC-Alkali-Rydberg-Calculator
pandas
numpy<2
scipy